java: final api client - preview. v1
add example project for java api

delete java example

add java example projectr

fix: update client

remove worker incomplete code, example and function that hides builder usage

remove function that hides builder usage

remove example

changing end of lines

remove extra code

remove extra code
