+++ packages/java/pom.xml
@@ -26,6 +26,15 @@
       <artifactId>grpc-netty</artifactId>
       <version>1.62.2</version>
     </dependency>
+
+    <!-- https://mvnrepository.com/artifact/io.projectreactor/reactor-core -->
+    <dependency>
+      <groupId>io.projectreactor</groupId>
+      <artifactId>reactor-core</artifactId>
+      <version>3.6.5</version>
+    </dependency>
+
+
     <dependency>
       <groupId>io.grpc</groupId>
       <artifactId>grpc-protobuf</artifactId>
