--- packages/java/src/main/java/armonik/client/event/spec/IEventClientAsync.java
@@ -1,10 +1,0 @@
-package armonik.client.event.spec;
-
-import armonik.client.event.impl.util.records.EventSubscriptionResponseRecord;
-import io.grpc.stub.StreamObserver;
-
-import java.util.List;
-
-public interface IEventClientAsync {
-  void waitForEventsUpdateBySessionId(String sessionId, List<String> resultIds, StreamObserver<EventSubscriptionResponseRecord> responseObserver);
-}
