--- packages/java/src/main/java/armonik/client/event/spec/IEventClientSync.java
@@ -1,10 +1,0 @@
-package armonik.client.event.spec;
-
-import armonik.client.event.impl.util.records.EventSubscriptionResponseRecord;
-
-import java.util.List;
-
-public interface IEventClientSync {
-  List<EventSubscriptionResponseRecord> getEventsUpdateBySessionId(String sessionId, List<String> resultIds);
-
-}
