--- packages/java/src/main/java/armonik/client/result/impl/util/records/DeleteResultsDataResponseRecord.java
@@ -1,6 +1,0 @@
-package armonik.client.result.impl.util.records;
-
-import java.util.List;
-
-public record DeleteResultsDataResponseRecord(String sessionId, List<String> deletedResultIds) {
-}
